StructuralEquationModes.jl: a fast and flexible SEM framework

This is a package for Structural Equation Modeling. It is still in development. Models you can fit include

Linear SEM that can be specified in RAM notation

ML, GLS and FIML estimation

Ridge Regularization

Multigroup SEM

Sums of arbitrary loss functions (everything the optimizer can handle)

We provide fast objective functions, gradients, and for some cases hessians as well as approximations thereof. As a user, you can easily define custom loss functions. For those, you can decide to provide analytical gradients or use finite difference approximation / automatic differentiation. You can choose to mix and match loss functions natively found in this package and those you provide. In such cases, you optimize over a sum of different objectives (e.g. ML + Ridge). This mix and match strategy also applies to gradients, where you may supply analytic gradients or opt for automatic differentiation or mix analytical and automatic differentiation.

You may consider using this package if:

you want to extend SEM (e.g. add a new objective function) and need an extendable framework

you want to extend SEM, and your implementation needs to be fast (because you want to do a simulation, for example)

you want to fit the same model(s) to many datasets (bootstrapping, simulation studies)

you are planning a study and would like to do power simulations

For examples on how to use the package, see the Tutorials.

Installation

To install the latest version from GitHub, use the following commands in your julia REPL:

using Pkg
Pkd.add("https://github.com/StructuralEquationModels/StructuralEquationModels.jl")

Citing the package

StructuralEquationModels.AbstractSem — Type

Most abstract supertype for all SEMs

StructuralEquationModels.SemLoss — Type

SemLoss(args...; ...)

Constructs the loss field of a SEM. Can contain multiple SemLossFunctions, the model is optimized over their sum. See also SemLossFunction.

Arguments

args...: Multiple SemLossFunctions.

Examples

my_ml_loss = SemML(...)
my_ridge_loss = SemRidge(...)
my_loss = SemLoss(SemML, SemRidge)

StructuralEquationModels.Sem — Type

Sem(;observed = SemObsCommon, imply = RAM, loss = (SemML,), diff = SemDiffOptim, kwargs...)

Constructor for the basic Sem type. All additional kwargs are passed down to the constructors for the fields.

Arguments

observed: isa SemObs or a constructor.

imply: isa SemImply or a constructor.

loss: Tuple of objects that are SemLossFunctions or constructors.

observed: isa SemObs or a constructor.

Returns a struct of type Sem with fields

observed::SemObs: Stores observed data, sample statistics, etc. See also SemObs.

imply::SemImply: Computes Σ, μ, etc. See also SemImply.

loss::SemLoss: Computes the objective and gradient of a sum of loss functions. See also SemLoss.

diff::SemDiff: Connects the model to the optimizer. See also SemDiff.

StructuralEquationModels.SemObs — Type

Supertype of all objects that can serve as the observed field of a SEM. Pre-processes data and computes sufficient statistics for example. If you have a special kind of data, e.g. ordinal data, you should implement a subtype of SemObs.

StructuralEquationModels.SemLossFunction — Type

Supertype for all loss functions of SEMs. If you want to implement a custom loss function, it should be of this type.

StructuralEquationModels.SemImply — Type

Supertype of all objects that can serve as the imply field of a SEM. Computed model-implied values that should be compared with the observed data to find parameter estimates, e. g. the model implied covariance or mean. If you would like to implement a different notation, e.g. LISREL, you should implement a subtype of SemImply.