Model fitting

As we saw in A first model, after you have build a model, you can fit it via

model_fit = sem_fit(model)

# output

Fitted Structural Equation Model 
=============================================== 
--------------------- Model ------------------- 

Structural Equation Model 
- Loss Functions 
   SemML
- Fields 
   observed:  SemObservedData 
   imply:     RAM 
   optimizer: SemOptimizerOptim 

------------- Optimization result ------------- 

 * Status: success

 * Candidate solution
    Final objective value:     2.120543e+01

 * Found with
    Algorithm:     L-BFGS

 * Convergence measures
    |x - x'|               = 6.13e-05 ≰ 1.5e-08
    |x - x'|/|x'|          = 8.21e-06 ≰ 0.0e+00
    |f(x) - f(x')|         = 1.05e-09 ≰ 0.0e+00
    |f(x) - f(x')|/|f(x')| = 4.94e-11 ≤ 1.0e-10
    |g(x)|                 = 2.48e-05 ≰ 1.0e-08

 * Work counters
    Seconds run:   0  (vs limit Inf)
    Iterations:    175
    f(x) calls:    524
    ∇f(x) calls:   524

You may optionally specify Starting values.

API - model fitting

StructuralEquationModels.sem_fit — Function

sem_fit(model::AbstractSem; start_val = start_val, kwargs...)

Return the fitted model.

Arguments

model: AbstractSem to fit

start_val: vector of starting values or function to compute starting values (1)

kwargs...: keyword arguments, passed to starting value functions

(1) available options are start_fabin3, start_simple and start_partable. For more information, we refer to the individual documentations and the online documentation on Starting values.

Examples

sem_fit(
    my_model; 
    start_val = start_simple,
    start_covariances_latent = 0.5)