Model inspection

After you fitted a model,

model_fit = fit(model)

you end up with an object of type SemFit.

You can get some more information about it by using the details function:

details(model_fit)

Fitted Structural Equation Model

--------------------------------- Properties ---------------------------------

Optimization algorithm:      L-BFGS
Convergence:                 true
No. iterations/evaluations:  127

Number of parameters:        31
Number of data samples:      75

----------------------------------- Model ------------------------------------

Structural Equation Model
- Loss Functions
   SemML
- Fields
   observed:    SemObservedData
   implied:     RAM

To compute fit measures, we use

fit_measures(model_fit)

Dict{Symbol, Union{Missing, Float64}} with 8 entries:
  :minus2ll => 3106.66
  :AIC      => 3168.66
  :BIC      => 3240.5
  :χ²       => 37.6169
  :dof      => 35.0
  :p_value  => 0.350262
  :nparams  => 31.0
  :RMSEA    => 0.0315739

or compute them individually:

AIC(model_fit)

3168.655882149098

A list of available Fit measures is at the end of this page.

To inspect the parameter estimates, we can update a ParameterTable object and call details on it:

update_estimate!(partable, model_fit)

details(partable)

---------------------------------- Variables ---------------------------------

Latent variables:    ind60 dem60 dem65
Observed variables:  x1 x2 x3 y1 y2 y3 y4 y5 y6 y7 y8

---------------------------- Parameter Estimates -----------------------------

Loadings:

ind60

  to   estimate   value_fixed   start   free   from    label   relation

  x1   1.0        1.0                   0.0    ind60   const   →
  x2   2.18                             1.0    ind60   θ_1     →
  x3   1.82                             1.0    ind60   θ_2     →

dem60

  to   estimate   value_fixed   start   free   from    label   relation

  y1   1.0        1.0                   0.0    dem60   const   →
  y2   1.26                             1.0    dem60   θ_3     →
  y3   1.06                             1.0    dem60   θ_4     →
  y4   1.26                             1.0    dem60   θ_5     →

dem65

  to   estimate   value_fixed   start   free   from    label   relation

  y5   1.0        1.0                   0.0    dem65   const   →
  y6   1.19                             1.0    dem65   θ_6     →
  y7   1.28                             1.0    dem65   θ_7     →
  y8   1.27                             1.0    dem65   θ_8     →

Directed Effects:

  from        to      estimate   value_fixed   start   free   label

  ind60   →   dem60   1.48                             1.0    θ_9
  dem60   →   dem65   0.84                             1.0    θ_10
  ind60   →   dem65   0.57                             1.0    θ_11

Variances:

  from        to      estimate   value_fixed   start   free   label

  x1      ↔   x1      0.08                             1.0    θ_12
  x2      ↔   x2      0.12                             1.0    θ_13
  x3      ↔   x3      0.47                             1.0    θ_14
  y1      ↔   y1      1.92                             1.0    θ_15
  y2      ↔   y2      7.47                             1.0    θ_16
  y3      ↔   y3      5.14                             1.0    θ_17
  y4      ↔   y4      3.19                             1.0    θ_18
  y5      ↔   y5      2.38                             1.0    θ_19
  y6      ↔   y6      5.02                             1.0    θ_20
  y7      ↔   y7      3.48                             1.0    θ_21
  y8      ↔   y8      3.3                              1.0    θ_22
  ind60   ↔   ind60   0.45                             1.0    θ_23
  dem60   ↔   dem60   4.01                             1.0    θ_24
  dem65   ↔   dem65   0.17                             1.0    θ_25

Covariances:

  from       to   estimate   value_fixed   start   free   label

  y1     ↔   y5   0.63                             1.0    θ_26
  y2     ↔   y4   1.33                             1.0    θ_27
  y2     ↔   y6   2.18                             1.0    θ_28
  y3     ↔   y7   0.81                             1.0    θ_29
  y8     ↔   y4   0.35                             1.0    θ_30
  y8     ↔   y6   1.37                             1.0    θ_31

We can also update the ParameterTable object with other information via update_partable!. For example, if we want to compare hessian-based and bootstrap-based standard errors, we may write

se_bs = se_bootstrap(model_fit; n_boot = 20)
se_he = se_hessian(model_fit)

update_partable!(partable, :se_hessian, model_fit, se_he)
update_partable!(partable, :se_bootstrap, model_fit, se_bs)

details(partable)

Number of nonconverged models: 0.0

---------------------------------- Variables ---------------------------------

Latent variables:    ind60 dem60 dem65
Observed variables:  x1 x2 x3 y1 y2 y3 y4 y5 y6 y7 y8

---------------------------- Parameter Estimates -----------------------------

Loadings:

ind60

  to   estimate   value_fixed   start   free   from    label   relation   se_bootstrap   se_hessian

  x1   1.0        1.0                   0.0    ind60   const   →          0.0            0.0
  x2   2.18                             1.0    ind60   θ_1     →          0.16           0.14
  x3   1.82                             1.0    ind60   θ_2     →          0.17           0.15

dem60

  to   estimate   value_fixed   start   free   from    label   relation   se_bootstrap   se_hessian

  y1   1.0        1.0                   0.0    dem60   const   →          0.0            0.0
  y2   1.26                             1.0    dem60   θ_3     →          0.14           0.19
  y3   1.06                             1.0    dem60   θ_4     →          0.17           0.15
  y4   1.26                             1.0    dem60   θ_5     →          0.12           0.15

dem65

  to   estimate   value_fixed   start   free   from    label   relation   se_bootstrap   se_hessian

  y5   1.0        1.0                   0.0    dem65   const   →          0.0            0.0
  y6   1.19                             1.0    dem65   θ_6     →          0.17           0.17
  y7   1.28                             1.0    dem65   θ_7     →          0.14           0.16
  y8   1.27                             1.0    dem65   θ_8     →          0.17           0.16

Directed Effects:

  from        to      estimate   value_fixed   start   free   label   se_bootstrap   se_hessian

  ind60   →   dem60   1.48                             1.0    θ_9     0.33           0.4
  dem60   →   dem65   0.84                             1.0    θ_10    0.09           0.1
  ind60   →   dem65   0.57                             1.0    θ_11    0.22           0.24

Variances:

  from        to      estimate   value_fixed   start   free   label   se_bootstrap   se_hessian

  x1      ↔   x1      0.08                             1.0    θ_12    0.01           0.02
  x2      ↔   x2      0.12                             1.0    θ_13    0.08           0.07
  x3      ↔   x3      0.47                             1.0    θ_14    0.07           0.09
  y1      ↔   y1      1.92                             1.0    θ_15    0.47           0.48
  y2      ↔   y2      7.47                             1.0    θ_16    1.16           1.37
  y3      ↔   y3      5.14                             1.0    θ_17    1.01           0.99
  y4      ↔   y4      3.19                             1.0    θ_18    0.79           0.77
  y5      ↔   y5      2.38                             1.0    θ_19    0.47           0.5
  y6      ↔   y6      5.02                             1.0    θ_20    0.96           0.91
  y7      ↔   y7      3.48                             1.0    θ_21    0.68           0.74
  y8      ↔   y8      3.3                              1.0    θ_22    1.03           0.72
  ind60   ↔   ind60   0.45                             1.0    θ_23    0.09           0.09
  dem60   ↔   dem60   4.01                             1.0    θ_24    0.91           0.96
  dem65   ↔   dem65   0.17                             1.0    θ_25    0.2            0.22

Covariances:

  from       to   estimate   value_fixed   start   free   label   se_bootstrap   se_hessian

  y1     ↔   y5   0.63                             1.0    θ_26    0.3            0.38
  y2     ↔   y4   1.33                             1.0    θ_27    0.68           0.71
  y2     ↔   y6   2.18                             1.0    θ_28    0.81           0.74
  y3     ↔   y7   0.81                             1.0    θ_29    0.44           0.63
  y8     ↔   y4   0.35                             1.0    θ_30    0.41           0.47
  y8     ↔   y6   1.37                             1.0    θ_31    0.8            0.58

Export results

You may convert a ParameterTable to a DataFrame and use the DataFrames package for further analysis (or to save it to your hard drive).

using DataFrames

parameters_df = DataFrame(partable)

34×10 DataFrame

Row	se_bootstrap	se_hessian	from	value_fixed	label	start	estimate	to	relation	free

Float64	Float64	Symbol	Float64	Symbol	Float64	Float64	Symbol	Symbol	Bool

1	0.0	0.0	ind60	1.0	const	NaN	1.0	x1	→	false

2	0.164162	0.139947	ind60	NaN	θ_1	NaN	2.18036	x2	→	true

3	0.170836	0.153213	ind60	NaN	θ_2	NaN	1.81849	x3	→	true

4	0.0	0.0	dem60	1.0	const	NaN	1.0	y1	→	false

5	0.144774	0.186748	dem60	NaN	θ_3	NaN	1.25683	y2	→	true

6	0.165231	0.149343	dem60	NaN	θ_4	NaN	1.0577	y3	→	true

7	0.120965	0.151747	dem60	NaN	θ_5	NaN	1.26484	y4	→	true

8	0.0	0.0	dem65	1.0	const	NaN	1.0	y5	→	false

9	0.174432	0.172499	dem65	NaN	θ_6	NaN	1.18573	y6	→	true

10	0.139531	0.161455	dem65	NaN	θ_7	NaN	1.27947	y7	→	true

11	0.167317	0.164313	dem65	NaN	θ_8	NaN	1.26596	y8	→	true

12	0.328655	0.399917	ind60	NaN	θ_9	NaN	1.48315	dem60	→	true

13	0.0869349	0.099454	dem60	NaN	θ_10	NaN	0.837368	dem65	→	true

14	0.222929	0.2353	ind60	NaN	θ_11	NaN	0.572317	dem65	→	true

15	0.0143915	0.0201154	x1	NaN	θ_12	NaN	0.0826479	x1	↔	true

16	0.0787025	0.0713897	x2	NaN	θ_13	NaN	0.121438	x2	↔	true

17	0.0658648	0.090994	x3	NaN	θ_14	NaN	0.473004	x3	↔	true

18	0.468997	0.478276	y1	NaN	θ_15	NaN	1.91706	y1	↔	true

19	1.16095	1.37271	y2	NaN	θ_16	NaN	7.47099	y2	↔	true

20	1.01307	0.988161	y3	NaN	θ_17	NaN	5.13669	y3	↔	true

21	0.790072	0.771056	y4	NaN	θ_18	NaN	3.18982	y4	↔	true

22	0.465031	0.498686	y5	NaN	θ_19	NaN	2.38278	y5	↔	true

23	0.95744	0.913521	y6	NaN	θ_20	NaN	5.02037	y6	↔	true

24	0.683692	0.742638	y7	NaN	θ_21	NaN	3.47799	y7	↔	true

25	1.0343	0.721323	y8	NaN	θ_22	NaN	3.29769	y8	↔	true

26	0.0947695	0.0885047	ind60	NaN	θ_23	NaN	0.454493	ind60	↔	true

27	0.914567	0.963669	dem60	NaN	θ_24	NaN	4.00902	dem60	↔	true

28	0.202752	0.224761	dem65	NaN	θ_25	NaN	0.174824	dem65	↔	true

29	0.29631	0.376502	y1	NaN	θ_26	NaN	0.632174	y5	↔	true

30	0.68354	0.712798	y2	NaN	θ_27	NaN	1.33007	y4	↔	true

31	0.813437	0.740916	y2	NaN	θ_28	NaN	2.18153	y6	↔	true

32	0.444655	0.633898	y3	NaN	θ_29	NaN	0.806343	y7	↔	true

33	0.407778	0.467127	y8	NaN	θ_30	NaN	0.352773	y4	↔	true

34	0.797636	0.58353	y8	NaN	θ_31	NaN	1.37423	y6	↔	true

API - model inspection

StructuralEquationModels.details — Function

(1) details(sem_fit::SemFit; show_fitmeasures = false)

(2) details(partable::AbstractParameterTable; ...)

Print information about (1) a fitted SEM or (2) a parameter table to stdout.

Extended help

Addition keyword arguments

digits = 2: controls precision of printed estimates, standard errors, etc.

color = :light_cyan: color of some parts of the printed output. Can be adjusted for readability.

secondary_color = :light_yellow

show_variables = true

show_columns = nothing: columns names to include in the output e.g.[:from, :to, :estimate])

StructuralEquationModels.update_estimate! — Function

update_estimate!(
    partable::AbstractParameterTable,
    fit::SemFit)

Write parameter estimates from fit to the :estimate column of partable

StructuralEquationModels.update_partable! — Function

(1) update_partable!(partable::AbstractParameterTable, column, fitted:SemFit, params, default = nothing)

(2) update_partable!(partable::AbstractParameterTable, column, param_labels::Vector{Symbol}, params, default = nothing)

Add a new column to a parameter table. column is the name of the column, params contains the values of the new column, and fitted or param_labels is used to match the values to the correct parameter labels. The default value is used if a parameter in partable does not occur in param_labels.

Additional functions

Additional functions that can be used to extract information from a SemFit object:

StructuralEquationModels.SemFit — Type

SemFit

Fitted structural equation model.

Interfaces

minimum(::SemFit) -> minimum objective value

solution(::SemFit) -> parameter estimates

start_val(::SemFit) -> starting values

model(::SemFit)

optimization_result(::SemFit)

optimizer(::SemFit) -> optimization algorithm

n_iterations(::SemFit) -> number of iterations

convergence(::SemFit) -> convergence properties

StatsAPI.params — Function

params(model)

Return all parameters of a model.

params(partable::ParameterTable) -> Vector{Symbol}

Return the vector of SEM model parameter identifiers.

StructuralEquationModels.param_labels — Function

param_labels(semobj) -> Vector{Symbol}

Return the vector of parameter labels (in the same order as params).

StructuralEquationModels.nparams — Function

nparams(semobj)

Return the number of parameters in a SEM model associated with semobj.

See also params.

StructuralEquationModels.nsamples — Function

nsamples(semobj)

Return the number of samples (observed data points).

For ensemble models, return the sum over all submodels.

StructuralEquationModels.nobserved_vars — Function

nobserved_vars(semobj)

Return the number of observed variables in a SEM model associated with semobj.

Fit measures

StructuralEquationModels.fit_measures — Function

fit_measures(sem_fit, args...)

Return a default set of fit measures or the fit measures passed as args....

StructuralEquationModels.AIC — Function

AIC(sem_fit::SemFit)

Return the akaike information criterion.

StructuralEquationModels.BIC — Function

BIC(sem_fit::SemFit)

Return the bayesian information criterion.

StructuralEquationModels.χ² — Function

χ²(sem_fit::SemFit)

Return the χ² value.

StatsAPI.dof — Function

dof(model::StatisticalModel)

Return the number of degrees of freedom consumed in the model, including when applicable the intercept and the distribution's dispersion parameter.

dof(sem_fit::SemFit)
dof(model::AbstractSem)

Return the degrees of freedom.

StructuralEquationModels.minus2ll — Function

minus2ll(sem_fit::SemFit)

Return the negative 2* log likelihood.

StructuralEquationModels.p_value — Function

p(sem_fit::SemFit)

Return the p value computed from the χ² test statistic.

StructuralEquationModels.RMSEA — Function

RMSEA(sem_fit::SemFit)

Return the RMSEA.